+++ R/read_au_ici.R
@@ -4,8 +4,6 @@
 #' @param path (Character) The path to the XLSX file.
 #' @param clean_non_ascii (Logical) Should the non-ASCII characters be
 #'   reasonably converted? Defaults to \code{FALSE}.
-#' @param version (Logical) Should the version information be included?
-#'   Defaults to \code{TRUE}.
 #' @details This function reads-in and automatically cleans the Australia
 #'   Industrial Chemicals Inventory.
 #' @return Returns a data frame.
@@ -31,18 +29,14 @@
 #'
 #' path <- "Full list of chemicals on the Inventory - 10 February 2022.xlsx"
 #'
-#' ici <- read_ici(path)
+#' ici <- read_au_ici(path)
 #' }
 #' @importFrom openxlsx read.xlsx
 #' @export
-read_au_ici <- function(path, clean_non_ascii = FALSE, version = TRUE) {
+read_au_ici <- function(path, clean_non_ascii = FALSE) {
 
   if (!is.logical(clean_non_ascii) || is.na(clean_non_ascii)) {
     clean_non_ascii <- FALSE
-  }
-
-  if (!is.logical(version) || is.na(version)) {
-    version <- TRUE
   }
 
   ici <- openxlsx::read.xlsx(xlsxFile = path, startRow = 2, cols = 1:5)
@@ -56,34 +50,18 @@ read_au_ici <- function(path, clean_non_ascii = FALSE, version = TRUE) {
     ici,
     cr_no = as.integer(cr_no),
     cas_no = ifelse(
-      test = cas_no == "Confidential",
+      test = cas_no == "Confidential" | !.check_cas(cas_no),
       yes = NA_character_,
       no = cas_no
     )
   )
 
   if (clean_non_ascii) {
-
     ici <- transform(
       ici,
       cas_name = .clean_non_ascii(cas_name),
       molecular_formula = .clean_non_ascii(molecular_formula)
     )
-
-  }
-
-  if (version) {
-    version <- unlist(strsplit(path, split = "/"))
-    version <- version[
-      sapply(
-        version,
-        FUN = function(x) { grepl(pattern = ".xlsx", x) },
-        USE.NAMES = FALSE
-      )
-    ]
-    version <- unlist(strsplit(version, split = "\\."))[1]
-    version <- unlist(strsplit(version, split = " - "))[2]
-    ici <- transform(ici, version = version)
   }
 
   ici